requirements.txt
Pillow==9.5.0
PyPDF2==3.0.1
zipfile36==0.1.3
rarfile==4.0
flask==2.3.3
flask-cors==4.0.0
tqdm==4.66.1
pyinstaller==6.0.0 